refactor: move each categor to it's own package
