+++ collector/processors/odigostailsamplingprocessor/processor.go
@@ -11,6 +11,9 @@ import (
 	"go.uber.org/zap"
 
 	"github.com/odigos-io/odigos/collector/processors/odigostailsamplingprocessor/category"
+	"github.com/odigos-io/odigos/collector/processors/odigostailsamplingprocessor/category/costreduction"
+	"github.com/odigos-io/odigos/collector/processors/odigostailsamplingprocessor/category/highlyrelevant"
+	"github.com/odigos-io/odigos/collector/processors/odigostailsamplingprocessor/category/noisy"
 	"github.com/odigos-io/odigos/collector/processors/odigostailsamplingprocessor/internal/metadata"
 	commonapisampling "github.com/odigos-io/odigos/common/api/sampling"
 	"github.com/odigos-io/odigos/common/collector"
@@ -65,7 +68,7 @@ func (p *tailSamplingProcessor) processTraces(ctx context.Context, td ptrace.Tra
 		}
 	}
 
-	matched, highlyRelevantOperationRule := category.EvaluateHighlyRelevantOperations(td, p.odigosConfigExtension)
+	matched, highlyRelevantOperationRule := highlyrelevant.Evaluate(td, p.odigosConfigExtension)
 	if matched {
 		percentageAtLeast := category.GetPercentageOrDefault100(highlyRelevantOperationRule.PercentageAtLeast)
 		keepTrace := tracePercentage <= percentageAtLeast
@@ -77,7 +80,7 @@ func (p *tailSamplingProcessor) processTraces(ctx context.Context, td ptrace.Tra
 		return ptrace.NewTraces(), nil
 	}
 
-	matched, costReductionRule := category.EvaluateCostReductionOperations(td, p.odigosConfigExtension)
+	matched, costReductionRule := costreduction.Evaluate(td, p.odigosConfigExtension)
 	if matched {
 		percentageAtMost := costReductionRule.PercentageAtMost
 		keepTrace := tracePercentage <= percentageAtMost
@@ -110,7 +113,7 @@ func (p *tailSamplingProcessor) evaluateNoisyOperations(td ptrace.Traces) (bool,
 		return false, nil
 	}
 
-	return category.EvaluateNoisyOperations(rootSpan, tailSamplingConfig.NoisyOperations)
+	return noisy.Evaluate(rootSpan, tailSamplingConfig.NoisyOperations)
 }
 
 func (p *tailSamplingProcessor) Start(ctx context.Context, host component.Host) error {
